client retry
Summary:
Currently if the shadow client creation fails then it is not retried (even though the client stays up). This led to SEV S169467. The solution proposed here is to periodically retry if appends fail.

Note that there are a few human intervention solutions already available:
1. Restarting the cluster restarts the client (heavy handed approach).
2. Updating shadowing information in the config will retry client creations (non-automatic).

I haven't yet tested it. Looking for feedback if this is a good approach.

Design:
There are two existing data structures:
1. client_init_queue_: queue of clients to be created
2. client_map_: map of clients already created.

We don't create something that is already there with the help of the map.

Currently the creation attempts go to the queue from where they are tried once. If it fails the context is thrown away -- there is no retry. In the new approach I created a map where re-try attempts are temporarily queued. 1 in N failed attempts add the context to the map. Every M seconds the ClientInitThread wakes up and drains the retry map into client_init_queue_ and then retries creations once more.

The failed ones are thrown away. This way we won't keep retrying something stale or unused. If they are needed then a subsequent appendShadow() will requeue it.

The client_map_ ensures that something directly added to the queue doesn't get created twice because it is also in the retry map.

Update:
- Added a setting to tune the retry interval. 0 disables the feature.
- Updated the integration test and made sure this actually works.
- Added stats for the retry mechanism.
- Add documentation for the feature.

Reviewed By: davidvigier

Differential Revision: D13553971

fbshipit-source-id: 0a51582202a35b8cda76adf981b72537a4b99679