Fix use after free in settings
Summary: SettingsEasyInit used to assume that the `const char* name` points to a constant string. It's a reasonable assumption, but (a) it's not documented (and it's not clear where to document it), (b) it's not true for RocksDBSettings (although I'm changing it in D13709088), (c) it probably doesn't win any noticeable amount of performance because SettingsEasyInit::operator() copies the `name` into an std::string, like, 5 times, and does lots of other heap allocations and slow stuff. So this diff removes this assumption and copies `name` into an std::string instead of retaining the original pointer.

Reviewed By: scsiguy

Differential Revision: D13709060

fbshipit-source-id: 15056dff198db99a1a37d6c41cec8bc629f115a5