Characterize remaining read/write params from ODS, Scuba
Summary: This diff makes ld-characterize-tier characterize the load on a cluster and produce a full read and write load specification for LDBench, such that only a few further steps are needed, as explained in the LogDevice testing wiki. To accomplish this, in addition to poking the cluster as added by D13147949, it now also queries ODS and Scuba.

Reviewed By: al13n321

Differential Revision: D13556561

fbshipit-source-id: 4714ebaaf3ae6e0a2f1167df2be89ac1a94c0070