Fix sequencer reactivation loop when esn-bits setting is small, and refactor the way sequencer handles config updates
Summary:
The test used to fail because it sets esn_bits to 8, which makes sliding window size to be clamped to 255, so SequencerBackgroundActivator kept saying that effective window size doesn't match the log attributes and reactivating the sequencer.

This diff refactors things a bit: window size is now extracted into a structure "things that should cause sequencer reactivation when they change", and this struct is stored in EpochSequencer unmodified (window size not clamped). Other things seemed like a good for for this structure: `esn_bits` setting, extras and synced_copies log attributes. Moving extras and synced_copies into the struct seems to have simplified code a bit (not as much as I hoped), but dropped the support for updating these things on the fly - now they cause sequencer reactivation; this seems fine given that we never change them ever.

Reviewed By: runemaster

Differential Revision: D13744843

fbshipit-source-id: 889f3b962a49a868dbc41d3da637ad7c1fe8cf6e