Re initialize the delta log readstream on re-enabling RSM
Summary:
If you disable the LCM and re-enable it again, delta log ReadStream doesn't get re-initialized. That's because when we re-enable it, we don't update the our sync_state_ to `SyncState::SYNC_SNAPSHOT` which means we don't enter the code path at  `ReplicatedStateMachine-inl.h:335` which re-enables the delta log readstream.

Let's recreate the whole RSM when LCM gets reenabled.

Honestly, I have no idea how this used to work before. Or maybe we didn’t exercise it before?

Reviewed By: davidvigier

Differential Revision: D13859732

fbshipit-source-id: 7f847f1351c009b155143bc7997ee42053ad313f