Fix the ConfigurationFetchRequest on OSS build
Summary: The test is failing on the OSS build with Status::INVALID because it's failing to create a connection to the node. Honestly it kind of makes sense and I'm not sure how it's working internally. To go around that, given that we are not actually testing the sending itself, let's use the Mock class instead and unmock the functions that we don't want to be mocked (timer functions).

Reviewed By: calebmarchent

Differential Revision: D13895634

fbshipit-source-id: 299f712f71c30f95879246e28ea6dd3faf639c6a