WorkContext to accept KeepAlive of underlying executor.
Summary:
KeepAlive makes sure underlying executor does not go out of scope.
Reasons for doing this :
1. `SerialExecutor` does not have a public constructor and provides a KeepAlive.
`SerializedWorkContext` based on `SerialExecutor` thus can pass only the KeepAlive
to WorkContext constructor hence it is needed.

2. Previous `WorkContext` constructor used `Executor` reference, which made ownership
semantics little difficult. `KeepAlive` provides a shared access to `Executor`
which solves the ownership problem. It allows `WorkContext` to not own the
`Executor` but at the same time provides guarantees that the executor does not go
away before the context.

This diff also introduces class `SerializedWorkContext` which uses the new constructor to construct the object.

Reviewed By: mcrnic

Differential Revision: D13777089

fbshipit-source-id: 161058b5f77a606bdf398d00934b46a57332515e