Fix unit test flakiness caused by unreliable waitUntilEventLogSynced()
Summary:
The "propagation" in "Propagated version" used to only include rebuilding coordinator, not per-worker authoritative status maps. I guess it was sufficient for the original use of that column, but not e.g. for how RebuildingTest.ReplicationCheckerDuringRebuilding uses it. RebuildingTest.ReplicationCheckerDuringRebuilding calls waitUntilEventLogSynced(), then creates a client and expects that this client will get an up-to-date shard status from servers. Without this diff this is often not the case, and very rarely it causes the test to fail.

This broke a test that was using "--event-log-grace-period" in a hacky way for an unrelated purpose. I added a separate setting for that purpose.

This also makes waitUntilEventLogSynced() wait for --event-log-grace-period, so I made all tests that use it set small --event-log-grace-period. Hope it's ok.

Reviewed By: davidvigier

Differential Revision: D13876754

fbshipit-source-id: 99e074f8400cfb981084656588076cd3a46d6957