Fix crash in RebuildingCoordinator when settings are updated shortly after aborting rebuilding
Summary: From logdeviced log: P60897553 - rebuilding restart is scheduled, but, before it happens, noteRebuildingSettingsChanged() calls trySlideGlobalWindow() with *latest* rebuilding set (that won't be applied to this RebuildingCoordinator until the scheduled rebuilding restart happens), and we fail an assert that checks that rebuilding set is in sync with RebuildingCoordinator's state. I introduced this bug in one of partition-by-partition diffs.

Reviewed By: adrienconrath

Differential Revision: D13934332

fbshipit-source-id: c51eb2740dc2fd2a55de4e9d7627404c8338f6bc