implement overwrite API
Summary:
NCM allows emergency tooling to overwrite the existing `NodesConfiguration`.
Note that we still require the versions of the NCs in the NCS to monotonically
increase. The first cut of the implementation returns an `E::VERSION_MISMATCH`
if the version number of the supplied config is not high enough.

In the future, NCM will encapsulate this complexity by automatically bumping
the version number and retry indefinitely until it succeeds.

Reviewed By: runemaster

Differential Revision: D13848543

fbshipit-source-id: 071290a548d871e9a857fe1fda6972e8dbec673a