implement NCM::update
Summary:
This diff implements the direct proposal method of NCM, which includes some
test util refactoring.

In addition, this diff adds the necessary edge case handling to treat an empty
string as a config with `EMPTY_VERSION` in `extractConfigVersion` and
deserialization. This allows an NCS to have an empty config as initial start up
state.

Reviewed By: runemaster

Differential Revision: D13823528

fbshipit-source-id: 56d3431bbfc9e53fabafd9e974668e7d2c6320b4