partition by partition: Progress estimation
Summary: Just a number between 0 and 1, in shard_rebuildings table. Hopefully it will increase at a more or less constant rate, and `SELECT avg(progress), min(progress) FROM ...` will be an ok estimate of how close rebuilding is to completing. (Probably won't be all that good because when some donors complete the rest go faster.)

Reviewed By: pvjykumar

Differential Revision: D13633293

fbshipit-source-id: 7da7c3ae472fc1f39584cc77f1502d70338bc9f9