Prevent iterator invalidation in processSocketsToClose
Summary: I suspect that the sockets_to_close_ vector may get modified during the closing of a socket causing the iterator to be invalidated and resulting in a crash. I think it's possible that while closing the socket and executing onSent callbacks with the reason, the callbacks may try to send another message immediately and cause another socket to be closed, hence pushing it back onto the same sockets_to_close_ vector that we are iterating over.

Reviewed By: mcrnic

Differential Revision: D14035083

fbshipit-source-id: 5478a4bf30f08fd59077eae4b69c8c41619c4aa8