Fix RebuildingTriggerQueueThresholdResetOnNonLeader test
Summary: This test was marked as flaky as sometimes it would fail because the stat rebuilding_supervisor_throttled on N1 was not unset (meaning the node had not exited throttling mode). This happens because of a race between when N1 sees N0 as started and the test checking stat. This diff fixes this by making sure we check for the stat only after we confirm that N1 has seen N0 as alive

Reviewed By: davidvigier

Differential Revision: D13708181

fbshipit-source-id: 9b53c3262fe5adb4a92812231e36abad7ef42e82