add util methods to detect and transition out of intermediary states
Summary:
Certain `StorageState`-s and `MetaDataStorageState`-s are considered
intermediary state in that NCM is reponsible to transition them to a
non-intermediary state (after a certain timeout / condition checks).

This diff adds the helper to be able to tell such states and to return the
destination state of such transition.

Reviewed By: MohamedBassem

Differential Revision: D14061259

fbshipit-source-id: 89d6ce2a6a109ae3bd0a67cf4a49b9245f7bb0ce