Deploy from CircleCI to DockerHub
In CircleCI we already compile LogDevice within a Docker image. For
this iteration, of the CI to DockerHub flow, we use a targeted Dockerfile
to pulls together the compiled artefacts into a publishable Docker image;
that users can download and run; following the builder/production image
structure implemented in Docker.ubuntu.

During testing, "Illegal Instruction" crashes were observed in using the
resulting container; to resolve this we apply the PORTABLE CMake option.
See https://forums.docker.com/t/getting-illegal-instruction-when-i-run-my-program-inside-container/43919

While in the future, it is planned to only create new Dockerimages on
successful builds, this limit is omitted from the current diff pending
resolution of the current failures.