EASY: explicitly initialize member variables
Summary:
In further testing, we found certain cases of uninitialized variables. It turns
out that the compilers do not do a very good job at identifying those. (MSAN is
designed to catch these bugs but unfortunately we don't have an MSAN platform.)

This diff explicitly initialize member variables in the membership / NC related
structs. It value-initializes POD types (numeric, enum, etc.) so that when the
owning class is default-initialized, they are not of indeterminate values.

Note that `map{}` or `unique_ptr{}` or `shared_ptr{}` is a no-op (the default
ctor would be invoked anyways if one does `unique_ptr<T> p;` since they are not
POD types, but I added them anyway to be future-proof and copy-pasta-safe.

Reviewed By: mcrnic

Differential Revision: D14150881

fbshipit-source-id: 5f1ab51643a3cbaf5d37e4aa469bfacaef992554