Python bindings for the NodesConfigurationAPI
Summary:
Implementing the first very basic python bindings for the NodesConfigAPI. The very first version includes: getting the NC, converting it into json, converting it back into binary and overwriting the config.

The `binary <-> json` conversion requires the FBS schema, which means that we will have to pack them in the XAR. I added python wrappers around finding and passing them to the c++ bindings to abstract the thing from the users of the module.

This module currently doesn't actually work because the NCM is not wired in the client yet. The logic itself works though.

Reviewed By: runemaster

Differential Revision: D13941860

fbshipit-source-id: b62e46cd478ea8983a5748def3e4615b33bb6e8e