tracker of per-shard last_changed_timestamp
Summary:
This diff introduces `ShardStateTracker` which keeps track of the per-shard
state for all shards that are currently in an intermediary (storage and
metadata storage) state. It tracks the first-known timestamp of the
transitions into the intermediary states and provides the ability to retrieve
shards that are known to have been in an intermediary state since before a
certain timestamp.

A follow-up diff will use the `ShardStateTracker` in the NCM to timeout
individual shards and propose changes to transition them out of the
intermediary states.

Reviewed By: MohamedBassem

Differential Revision: D14035627

fbshipit-source-id: 538f87dde32458874811ec22053329ec67526de4