Client support for custom stats + a few LDBench stats
Summary: We've got many services or tools, like replication checker or LDBench, that have a LogDevice client. In order that these may push stats without reinventing the wheel, this diff makes the LogDevice client capable of pushing custom stats sets for them, along with the regular client stats. For example, this diff makes LDBench create and register a stats holder for its custom stats. Those stats are defined in a '.inc' file, and the stats holder is told to include specifically those stats. The client's stats collection thread will then publish those stats along with the client's own stats, and any other sets of stats that are registered with the client. Long-term, StatsHolder will be refactored to make the code cleaner, see T40895127. A more well thought through set of stats for LDBench workers will be shipped in a subsequent diff.

Reviewed By: al13n321

Differential Revision: D13999513

fbshipit-source-id: 2d190e3c82a226f046c8cc15569738d41dedc3a4