Undeprecate STORE_Header::RECOVERY
Summary:
At first I tried to remove this "deprecated" flag as the comments suggested, but then found out that the comments are lying. Fixing the comments.

It *is* possible to get rid of this flag by assuming that `WRITTEN_BY_RECOVERY | REBUILDING` always means that the STORE message originates from a rebuilding, not from recovery. This is kind of supposed to be true since rebuilding doesn't run on unreleased records. But (a) there are probably corner cases where this can happen, (b) meta-fixer (and maybe scrubber in future?) set RECOVERY flag in STORE messages to differentiate themselves from append stores and to route replies correctly; those STOREs can legitimately be for records previously written by rebuilding, and setting REBUILDING flag in STORE causes the storage node to set WRITTEN_BY_REBUILDING flag in local log store, which seems potentially useful.

Reviewed By: runemaster

Differential Revision: D14238934

fbshipit-source-id: 0d83cdf4e32250ac99704bf2ebcd6bfa574525df