Fix how node is looked up in initServerSocket()
Summary: NodeID is composed of generation number and node index. In initiServerSocket(), we currently lookup node config using only the index. This is error prone; the node config should be looked up by using the entire NodeID to ensure that an error is returned when a NodeID containing valid index but invalid generation number is provided.

Differential Revision: D14180159

fbshipit-source-id: e24e2d32f2318ebc56d278e736f38943e64c270e