Fix some difference between record and CSI merge operators
Summary:
They used to treat WRITTEN_BY_{REBUILDING,RECOVERY} flags differently: for records, it'd bitwise OR the flags of highest-pri full record and highest-pri amend; for csi it would just take the flags from highest-pri entry. The latter is particularly incorrect for rebuilding amends because FLAG_WRITTEN_BY_REBUILDING is specifically _not_ set for amends even if they're written by rebuilding; so if the last merge operand is amend, csi won't have the flag, while record might.

This diff makes both record and csi merges bitwise OR the WRITTEN_BY_{REBUILDING,RECOVERY} flags of _all_ operands.

CSI_FLAG_HOLE is still problematic: for records it's taken from the full record, not from the amend, while for csi entries we can't do that because we don't know which entries correspond to full records. Taking the flag from the amend for records would probably be even worse because then we may end up with a record with empty payload but no HOLE flag, and it may end up being delivered to the client. I guess the solution would be to add CSI_FLAG_AMEND; then we can probably make the record and csi merging logic exactly the same, maybe even reuse some code. That's not in this diff.

The other csi flags are fine: DRAINED is always taken from the last operand, SHARD_ID doesn't need to be in sync between data and csi.

Reviewed By: scsiguy

Differential Revision: D13619826

fbshipit-source-id: c4f8ef4a57a806d2a5f0dd245062b8907f71242c