Fix for setting dscp
Summary:
Server needs access to client_dscp in order to set it for incomming connection.
If after negotiation we figure out other side is another node we need to update dscp.
This is hacky but this whole thing is getting reorganised so it does not matter much at this point

Reviewed By: tau0

Differential Revision: D14420219

fbshipit-source-id: afd2460300069b1335574952f4a2177f96cf2e24