by default, wait till initialization is finished before returning
Summary:
`NCM::init` has an async component which is done on the NCM worker thread (via
InitRequest). Currently `NCM::init()` does not wait for that request to get
processed before returning. This has caused confusion / races in client /
tooling usages where the short-lived client destroyed NCM before the init was
finished.

This diff makes wait_till_initialized the default behavior, assuming `init` is
called on a different thread. It also refactors so that the init_nc is passed
in instead of read through the Processor in deps.

Reviewed By: MohamedBassem

Differential Revision: D14410866

fbshipit-source-id: e9ca09f820ce1e1d035d2a6233ae41acaf2635f5