Removing all required fields in admin API
Summary:
I'm removing all required fields in Admin API according to the recommendation (https://our.intern.facebook.com/intern/wiki/Idl/optional-required-fields/)

Quoting from the documentation
```
it is often best to leave a field as unspecified (neither 'required' nor 'optional') unless an object
is expected to be changed and re-serialized. This is because both required and optional are
susceptible to errors of omission (i.e forgetfulness) on the part of developers.
```

My migration plan is to:
1- Update all tooling first (mainly ldshell, featureWiki, FBJE, alert_inspector). All of these are built with contbuild and deployed automatically
2- Build a new admin_server.

Reviewed By: MohamedBassem, ylectric

Differential Revision: D14502198

fbshipit-source-id: 142d38df470c3ea3134132cf3a33e59b01bd64af