StorageSetAccessor: clean up debug info stuff, add trace collection, print the trace in Mutator
Summary:
Before this diff, StorageSetAccessor had 3 (!) similar ways of reporting debug info:
 - printShardStatus(); prints authoritative statuses, states and statuses; omits shards in SUCCESS state; used by findKey,
 - getDebugInfo(); prints information about current wave, and all shard states and statuses; used when mutation times out,
 - allShardsStateSummary(); prints all shard states and statuses, and which shards are required; used for some error logging in StorageSetAccessor itself, and in a few requests.

This diff merges them all in one method describeState(), which tries to be clever and include only the interesting information. Just read the code.

Also added a sort of event tracing, see getDebugTrace() and ADD_TO_TRACE. It's used for dumping a detailed report of what each Mutator did. Would have been useful for T40989544, although I figured it out without this. Not sure if this tracing is useful for anything apart from Mutator - usually we're only interested in the final state of all shards, not in how we got there.

Reviewed By: gdavidsson

Differential Revision: D14255724

fbshipit-source-id: 964f034af0847de3fc28f257022af3c81ea5334a