Fix NodeStatsControllerIntegrationTest/NodeStatsControllerIntegrationTest.AdaptiveBoycottDuration/1 flakiness
Summary:
Copy-pasting a comment from the task:

It fails when one random append (out of many) fails, which causes an innocent node to get boycotted, which delays the test by a lot, waiting for it to get unboycotted again. This sounds bad on so many levels:
 1. Append timeout is set so low that one store timeout causes append failure.
 2. Boycotting doesn't tolerate even a single append failure (with 30 successes). Or maybe that's good?
 3. The outlier detection algorithm is complicated enough that it's hard to tell from the settings whether 1 failure out of 30 appends causes boycotting. 1 out of 100? out of 1000? All you know is that --node-stats-boycott-required-std-from-mean is 1, and a few other similarly obscure parameters. Not obvious how to improve it though.
 4. I couldn't figure out why the stores timed out. But I'd prefer the test to not require every single store (thousands of them) to be processed within 50ms.
 5. The spurious boycotting takes so long to expire that the test times out. Or maybe that's good, and the test can be made stricter and outright fail if an unexpected node gets boycotted? but then it needs to be reconciled with 2 or 1.

[End of copy-pasta.]

This diff fixes the flakiness by adding `--node-stats-boycott-sensitivity=0.03` to avoid boycotting nodes with <3% absolute failure rate.

Also decreases store timeout from 50ms to 10ms for all integration tests. That alone wasn't enough to avoid flakiness of this tests, appends would still occasionally time out 3 waves. But I'm keeping it anyway because it seems good to have integration tests use the same initial timeout as the logdeviced default.

Reviewed By: MohamedBassem

Differential Revision: D14452131

fbshipit-source-id: 0795977332e4fd760cb67365fe30c6e6c69f870b