SocketDependencies to accept Processor and Sender instance during
Summary:
SocketDependencies provide parent references to socket to fetch stats, config,
and access various Sender API's.

Miro is planning to rewrite Socket but untill that happens in order to unblock
EventLoop and Worker separation we need to remove all the Worker::onThisThread
references in SocketDependencies.

As part of this diff I am targeting just Processor and sender next set of diff
will handle other things. I am not changing Socket constructor right now
because it is going to change anyway later. But it made sense to make Socket
depend on Processor and Sender as in the new model also those things might
remain the same.

This should work now and after EventLoop and Worker separation because socket
is always created within Worker context either as part of NewConnectionRequest
or during sendingMessage to someone which does not have a valid socket.

Reviewed By: mcrnic

Differential Revision: D14094706

fbshipit-source-id: e5f1afd5456e01095f1b4a7a634a9903c65ef01d