Use common CMake modules to find glog and gflags
Glog behaviour does not change, just different variable names used.
GFlags changes from module to config target building on Ubuntu Bionic;
this has the cascading effect of switching logdevice from using a real
library name to an imported target "gflags_shared" which in-turn changes
the inclusion order such that we get link errors for glog looking for
gflags. To resolve this issue, gflags is added to the external deps and
we need to find it in the system before glog.

Test Plan:

Build LogDevice locally and with CircleCI.