Use common CMake modules to find glog and gflags (#72)
Summary:
Glog behaviour does not change, just different variable names used.
GFlags changes from module to config target building on Ubuntu Bionic;
this has the cascading effect of switching logdevice from using a real
library name to an imported target "gflags_shared" which in-turn changes
the library link order such that we get link errors for glog looking for
gflags. To resolve this issue, gflags is added to the external deps and
we need to find it in the system before glog.
Pull Request resolved: https://github.com/facebookincubator/LogDevice/pull/72

Reviewed By: Zepheus

Differential Revision: D14619245

Pulled By: calebmarchent

fbshipit-source-id: a5d293e816afed1520d57639134227d02b82f80f