ClusterMaintenanceWrapper wrapper around ClusterMaintenanceState
Summary:
This is a wrapper class around thrift's `ClusterMaintenanceState` that indexes the definitions and offer easy access to the target states for shards and sequencers.

This is designed with the maintenance manager requirements in mind. The RSM will wrap the ClusterMaintenanceState in this class.

This also adds the `force_restore_rebuilding` in the maintenance definition.

Reviewed By: pvjykumar

Differential Revision: D14612446

fbshipit-source-id: 9d9fb81946e0b689dfe0b57a14a8195a20bfe37b