Using Membership StorageState and MetaDataStorageState in Admin API
Summary:
This deprecates our custom ShardStorageState type in AdminAPI in favor of including the cluster membership's own StorageState. It also adds a new MetaDataStorageState to show which shards are metadata.

In future diffs, I'll move ldshell and other tools to use these new fields then once this is deployed everywhere we can remove the deprecated fields.

Reviewed By: calebmarchent

Differential Revision: D14642393

fbshipit-source-id: 451d49df860f220b8fb4024df7a8ad72a4c4274e