Fix shadow tests
Summary:
D14485613 lets the client evaluate logsconfig from the local cache, so
when we do any logsconfig testing, we'll need to manually sync, in order not to
be flaky. Currently, we only do this in shadow_test, but I put the wrapper
in test_helpers in case we expand.

Also updated the .pyi file so that our type signatures are properly inferred by
python.

Reviewed By: ylectric

Differential Revision: D14677880

fbshipit-source-id: e2dc5e534df1d9f39065895683ea126a92adfc61