Do not trigger purging on release message if log does not exist
Summary: Release message for a log that is no longer in config could cause purging to be run again and again even if it will never succeed (since log is no longer in config). This diff tries to fix this case by dropping the release if the log id is not in the config.

Reviewed By: runemaster

Differential Revision: D14588194

fbshipit-source-id: 535d055edad1f96284e939e687cf018fd855cc01