Make SocketDependencies independent of worker - SSLFetcher (3/n)
Summary:
Instead of having per worker SSLFetcher, associate it to a processor. There was
no particular reason why SSLFetcher was on a Worker, other than thread safe
update procedure.

This diff adds a mutex so that the context can be updated safely.

With enough in-region and out-region sockets we might see the context getting loaded frequently. Because, in-region uses null ciphers while out-region does not. I feel it should be ok as this was a problem today as well. Only thing that cannot happen now is parallel load of SSLContexts.

Reviewed By: mcrnic

Differential Revision: D14145535

fbshipit-source-id: 45fe9e8cd7070ae0984c6f7b9227e4445491b93f