Remove Worker references from ZeroCopiedDisposal
Summary:
Following changes removed dependence of ZeroCopiedRecord on ZeroCopiedDisposal. That helps in removing disposal logic completely from worker. Also disposal itself does not have to depend on any threadlocals.

**Usage of ZeroCopiedRecordDisposal**
Used by `EpochCacheEntry`  to release `ev_buffer_` allocated on some eventloop to be deleted on the same eventloop thread.
Rest of the usage of ZeroCopiedRecord already delete the record on the same thread, on which the record gets created. For example, Appender. But, anyways this implementation does not depend on such instances. The Disposal method explained below is generic and implicit. This makes sure right action is taken every time.
`ZeroCopiedRecordDisposal` used to batch deleted ZeroCopiedRecords and delete on an **explicit drain** from Worker/EventLoop thread.

**ZeroCopiedRecord and PayloadHolder**
If we look carefully, `ZeroCopiedRecord` itself can be deleted on any thread. `PayloadHolder` which contains the ev_buffer_ actually has the requirement to be deleted on the same thread. Hence, the custom deleter should be on the `ev_buffer_` rather than entire `ZeroCopiedRecord`.

**Custom Fitting**
`ZeroCopiedRecordDisposal` depended too much on ZeroCopiedRecord. If someone else wanted the same functionality, it could not be used as it was custom fit for just ZeroCopiedRecords.

**Ideally what we want**
1. Ability to create `shared_ptr<struct evbuffer>` by zero copy.
2. If `PayloadHolder` has a zero copied event_buffer delete it on the same thread which allocated that event_buffer.
3. Provide batched delete functionality like ZeroCopiedDisposal.
4. Avoid the need for explicit drain to drain the records.

**How is this achieved ?**
1. Added a class `ZeroCopyPayload` that wraps `struct evbuffer`. The constructor is hidden and object can be constructed only via static create method of `ZeroCopyPayload`.
2. Added a class `ZeroCopyDisposal` that is a custom deleter for ZeroCopyPayload. ZeroCopyPayload::create, sets the deleter as ZeroCopyDisposal and returns a shared_ptr of ZeroCopyPayload . ZeroCopyDisposal also has EventLoop instance , it is used when a ZeroCopyPayload  is deleted in the following manner.
3. `EventLoop` provides dispose method and also contains an instance of `BatchedBufferDisposer<ZeroCopyPayload>` . On invocation of EventLoop::dispose the implementation invokes dispose on `BatchedBufferDisposer`.
4. `BatchedBufferDisposer` is a class that accepts an Executor. It contains a list of records to be deleted and an atomic to get status of enqueued cleanup function atomically. Check out the implementation for more details.

Reviewed By: mcrnic

Differential Revision: D13969661

fbshipit-source-id: 96cd59ab3428e0201b5bdabbc7814ae9ee686f34