Make storage response to not use reply_pump. Make WheelTimers use the same hi priority mechanism
Summary:
RequestPump was used to reply back to response from StorageThreads to Worker Threads. The receive end of RequestPump queue uses Worker::onThisThread and assumes that  it runs on Worker threads. This is not true once we separate EventLoop and Worker.

Following options were considered :
**Post to Worker**
Instead of having this separate RequestPump for storage tasks the responses could be posted directly into the worker because it is also an Executor. This was tried previously but it lead to significant drop in performance.

**Create SerialExecutor**
Creating another SerialExecutor in the Worker will make the worker execute in parallel in future once Worker can execute on any threads. Worker is supposed to execute in single threaded context.

**High Pri Queue**
Adding another high priority queue within worker which picks up and  executes high pri task within a worker whenever any task within worker is  scheduled. Worker is a executor hence it can pick up new tasks as it wishes to execute.

The last idea is implemented in this diff. This should provide us the necessary performance and also not lag some performance critical applications.

I have changed the Timer to use this as well.

**Implementation Details**
I also add a low priority queue in the worker. These two queues would be moved to a SerialPrioritizedExecutor later. For now, when a task is enquequed the actual task is pushed to the appropriate queue based on it's priority. To the underlying executor we post a generic function which basically dequeues from one of the queues when the executor schedules the task.

The logic to select queue is based on discrete distribution where high priority queue is given weight of 70 and low priority queue is weighted at 30. In absence of work in one of the queues other priority level gets a chance to run. This makes sure that we do not starve or skip tasks indefintely. This mechanism helps us to add another priority level and keeps number of times a priority level is serviced bounded.

I will like to add tests to make sure dequeue is guaranteed . Also, it will be nice to have the distribution or weights as settings instead of hard coded values. I will do both of these once the discrete distribution idea is ok with reviewers.

Reviewed By: mcrnic

Differential Revision: D14197216

fbshipit-source-id: d51551911d48be6852f59c81f162731e44bc0da5