Remove trim method in Socket.
Summary:
trim method was leftover after lots of refactoring and did not have any use case except in close. Did not make sense to have  a separate interface for it. Removing it as part of this diff.

In case of close trim was invoking onSent for messages in DEFERRED mode, after that it used to make sure to deliver it anyways in the next statement. Did not see a reason to have it that way. Now onSent is called inline when trimming the pending queue.

Reviewed By: mcrnic

Differential Revision: D14715780

fbshipit-source-id: af31c969fdb4a4f95b54cec6781a215bbd91765b