Refactor Socket::receiveMessage
Summary:
Socket::receiveMessage had a lot of code which used to process message header and body. As part of this diff I am breaking receiveMessage into smaller parts especially the processing of the message body.

Processing of the message body itself is now divided in somewhat logical parts with code cleanup and avoiding code duplication especially when it comes to processing handshake message.

Once we have this we can go forward and complete processing of serialization and deserialization on separate thread.

Reviewed By: mcrnic

Differential Revision: D14718485

fbshipit-source-id: 48dfd2d5aa3c5bad82f43b4d7e7af61d15a01d3b