Make SocketEvent methods virtual
Summary:
Connection and Socket will have inheritance relationship for now till we transistion completely to new Sockets. We decided to do this as Socket API today will be migrated to Connection. Some of the methods will be made async but we prefer to do that transistion one at at time so that we always can make progress in small chunks.

Socket callbacks accesses worker pointer. But once Worker and EventLoop are separated, EventLoop::onThisThread won't get the Worker.  In order to allow current code to work, and to transition to new model slowly we still need a way to handle Socket callback in the current way i.e. by doing Worker::onThisThread().

My plan is,  override SocketEvent apis in Connection and setup worker pointer so the Worker::onThisThread calls get the right pointer before calling the usual Socket api. This way I am able to separate Worker and EventLoop and also allow current code to access Worker pointer.

Reviewed By: mcrnic

Differential Revision: D14719088

fbshipit-source-id: 216ff9049ebe3a2a11befa13fb01dbea34ad31ce