Add save and restore of context to EventLoop and SerialWorkContext
Summary:
Actual usage is in a later diff. Idea is to save thread locals like Worker pointer in RequestContext and save the RequestContext shared_ptr in the task posted to the executor. When work posted to worker is scheduled for execution,  restore the requestcontext captured in the lambda, it so that task under execution can access it.

Worker::onThisThread will be changed to fetch an Worker pointer from the RequestContext currently loaded.

Reviewed By: mcrnic

Differential Revision: D14274763

fbshipit-source-id: e33573f733424ec36511efbd807f478c53ac74ca