Fix crash in GetEpochRecoveryMetadataRequest when a node changes from disabled to read-write
Summary:
Had this crash on two nodes.

I'd guess that there are other places with the same bug, but even that is probably just the tip of the iceberg of how sloppily nodes config changes are handled everywhere. So I'm only fixing the crash that I saw and not going into the rabbit hole of trying to fix handling of nodes config everywhere.

Reviewed By: pvjykumar

Differential Revision: D14787604

fbshipit-source-id: 5db48cd1e6bc2b88ba10c01a26111a2b734edc10