Separate Worker from EventLoop.
Summary:
This diff contains contents of D14289438, D14739096, D14289962, D14762890,
D14290809, D14762873, D14291148, and D14312731.

Remove assumption that Worker can run on a EventLoop. Worker::onThisThread
depends on the thread local and ability to downcast EventLoop to Worker. But in
the new model that assumption is no longer true.

Diff D13807065 made sure that all requests posted go through the Worker.
Diff D14274763 made sure that when a request is posted to EventLoop or to
SerialWorkContext context is saved and restore when the task is scheduled for
execution.

Using those diffs, in here I am adding setting work context when anyone invokes
Worker::add or Worker::addWithPriority. The Executor take care of saving and
restoring this context.

Also in this diff I fix Worker::onThisThread to always load from context.

LifeCycle of new worker.

This diff includes meat of the logic related to separating worker from
eventloop and introducing it to new relationship with WorkContext.

This diff changes createWorker to separate event loop creation from worker
construction.
Worker constructor now accepts keep alive of the executor on which it is going
to execute.
Destruction of the worker and eventLoop is incomplete in the diff. Diff
D14291148 takes care of destroying EventLoop instances. There is some more work
to be done in order to destroy worker itself because of shutdown messages sent
during destruction.

Shutdown changes with new Worker model.

Worker takes a keepalive of EventLoop hence it was no longer possible to
delete eventloop before the thread exited.

As part of this diff eventLoop instance is deleted in one place irrespective of
how and why it was created.

Shutdown with new model still needs some work to support old code. Old code
like ClientReadStreams send message as part of destruction. But as deletion now
happens on main thread which does not have a event base sending message message is
not possible. Added a temporary fix by which ClientReadStream does not send a
shutdown message when bringing down the system.

Reviewed By: mcrnic

Differential Revision: D14811667

fbshipit-source-id: 9914379aa0190775612d846730e618cc3e34fa15